mcp-server/requirements.txt
mcp[cli]>=1.2.0
httpx>=0.27.0
google-cloud-run>=0.10.0
google-cloud-build>=3.20.0
google-cloud-billing>=1.13.0
google-cloud-billing-budgets>=1.14.0
uvicorn>=0.30.0
